src/data/usecases/add-account/db-add-account.ts
import {
  AccountModel,
  AddAccount,
  AddAccountModel,
  AddAccountRepository,
  Encrypter,
} from "./db-add-account-protocols";

export class DbAddAccount implements AddAccount {
  private readonly encrypter: Encrypter;
  private readonly addAccountRepository: AddAccountRepository;

  constructor(
    encrypter: Encrypter,
    addAccountRepository: AddAccountRepository
  ) {
    this.encrypter = encrypter;
    this.addAccountRepository = addAccountRepository;
  }

  async add(accountData: AddAccountModel): Promise<AccountModel> {
    const { name, email, password } = accountData;
    const hashedPassword = await this.encrypter.encrypt(password);
    const account = await this.addAccountRepository.add({
      name,
      email,
      password: hashedPassword,
    });
    return account;
  }
}
